requirements.txt
# GUI
PyQt6

# Database
SQLAlchemy
alembic

# API Clients
spotipy
plexapi>=4.15.4
anthropic>=0.5.0

# Development Tools
python-dotenv
pytest
black
flake8

# Utilities
aiohttp
requests>=2.31.0
tenacity>=8.2.3

# Date Handling
python-dateutil>=2.8.2